requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
asyncio
apscheduler
python-multipart
orjson
python-json-logger
psutil
pytest
pytest-asyncio
httpx
black
flake8
mypy
gunicorn
